requirements.txt
agno>=0.1.0
groq>=0.11.0
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
yfinance>=0.2.0
duckduckgo-search>=6.0.0
